intimg: address comments
